refactor: folder structures
