+++ packages/react-resval/__test__/system.spec.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { setBreakpoints, extendsBreakpoints, sortBreakpointsTrack, trackBreakpoints } from '../src/system'
-import { DefaultBreakpoints } from '../src/constants'
+import { DefaultBreakpoints } from '../src/v0.1.0/constants'
 
 describe('Check the `setBreakpoints` utility so that it gives the expected output.', () => {
   test('should return default breakpoints values', () => {
